# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

/**
 * Arbitrary client-state JSON snapshot (assessment history, bookmarks, etc.)
 */
export type CreateBackupRequestData = { [key: string]: unknown } | unknown[];

export interface CreateBackupRequest {
  /** Optional existing recovery code. If omitted, a new code is generated. */
  code?: string | null;
  /** Arbitrary client-state JSON snapshot (assessment history, bookmarks, etc.) */
  data: CreateBackupRequestData;
  /**
   * Optional schema version of the local data shape.
   * @minimum 1
   */
  version?: number;
}

export interface BackupSummary {
  code: string;
  version: number;
  updatedAt: string;
  createdAt: string;
}

export type RestoreBackupResponseData = { [key: string]: unknown } | unknown[];

export interface RestoreBackupResponse {
  code: string;
  version: number;
  data: RestoreBackupResponseData;
  updatedAt: string;
  createdAt: string;
}