# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: backup
    description: Numoo cloud backup & restore
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /backup:
    post:
      operationId: createBackup
      tags: [backup]
      summary: Create or update a Numoo cloud backup
      description: |
        Saves the family's local Numoo state under a new short recovery code,
        OR updates an existing backup if a code is provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBackupRequest"
      responses:
        "200":
          description: Backup created/updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupSummary"
        "404":
          description: Code provided does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /backup/{code}:
    get:
      operationId: restoreBackup
      tags: [backup]
      summary: Restore a Numoo backup by code
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backup restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreBackupResponse"
        "404":
          description: Code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    CreateBackupRequest:
      type: object
      properties:
        code:
          type: string
          description: Optional existing recovery code. If omitted, a new code is generated.
          nullable: true
        data:
          description: Arbitrary client-state JSON snapshot (assessment history, bookmarks, etc.)
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
        version:
          type: integer
          description: Optional schema version of the local data shape.
          minimum: 1
      required:
        - data
    BackupSummary:
      type: object
      properties:
        code:
          type: string
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - code
        - version
        - updatedAt
        - createdAt
    RestoreBackupResponse:
      type: object
      properties:
        code:
          type: string
        version:
          type: integer
        data:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - code
        - version
        - data
        - updatedAt
        - createdAt