# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Saves the family's local Numoo state under a new short recovery code,
OR updates an existing backup if a code is provided.

 * @summary Create or update a Numoo cloud backup
 */

export const CreateBackupBody = zod.object({
  code: zod
    .string()
    .nullish()
    .describe(
      "Optional existing recovery code. If omitted, a new code is generated.",
    ),
  data: zod
    .union([zod.record(zod.string(), zod.unknown()), zod.array(zod.unknown())])
    .describe(
      "Arbitrary client-state JSON snapshot (assessment history, bookmarks, etc.)",
    ),
  version: zod
    .number()
    .min(1)
    .optional()
    .describe("Optional schema version of the local data shape."),
});

export const CreateBackupResponse = zod.object({
  code: zod.string(),
  version: zod.number(),
  updatedAt: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Restore a Numoo backup by code
 */
export const RestoreBackupParams = zod.object({
  code: zod.coerce.string(),
});

export const RestoreBackupResponse = zod.object({
  code: zod.string(),
  version: zod.number(),
  data: zod.union([
    zod.record(zod.string(), zod.unknown()),
    zod.array(zod.unknown()),
  ]),
  updatedAt: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});