# lib/api-zod/src/generated/types/createBackupRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBackupRequestData } from "./createBackupRequestData";

export interface CreateBackupRequest {
  /** Optional existing recovery code. If omitted, a new code is generated. */
  code?: string | null;
  /** Arbitrary client-state JSON snapshot (assessment history, bookmarks, etc.) */
  data: CreateBackupRequestData;
  /**
   * Optional schema version of the local data shape.
   * @minimum 1
   */
  version?: number;
}