# lib/api-zod/src/generated/types/restoreBackupResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RestoreBackupResponseData } from "./restoreBackupResponseData";

export interface RestoreBackupResponse {
  code: string;
  version: number;
  data: RestoreBackupResponseData;
  updatedAt: Date;
  createdAt: Date;
}